+++ AtoCash/Models/DisbursementsAndClaimsMasterDTO.cs
@@ -19,7 +19,7 @@ namespace AtoCash.Models
         public int? SubProjectId { get; set; }
         public int? WorkTaskId { get; set; }
         public DateTime RecordDate { get; set; }
-        public decimal Amount { get; set; }
+        public Double Amount { get; set; }
         public int CostCentreId { get; set; }
         public int ApprovalStatusId { get; set; }
